ternaryOperator
A function that takes a string as input and, if the first letter is not capitalized, returns the reversed version of the original string. If the first letter is capitalized, then the string is returned unchanged. If an empty string is passed as input, the function must return an empty string.